Prefer /boot/coreos only if /boot/coreos/vmlinux-* exists
In case a user or old software created the first_boot flag file under
/boot/coreos instead of /boot/flatcar, the update breaks because /boot/coreos
exists and is prefered ofer /boot/flatcar when the kernel is written.
Make updating more robust by checking if the kernel binaries exist which as least
allows the user to create /boot/coreos/first_boot without breaking the system.