Report rollbacks as Omaha error codes
When the system failed to boot into an update, a rollback to the
previous partition took place. However, an update success was reported
via Omaha.
Add a new error code that is reported for rollbacks. The detection is
done by comparing the currently running version number with the
previous version number that is already stored for reportting also
serves as a way to detect if a reboot is needed.