Support optional machine alias besides the machine ID
The machine ID is the unique identifier of a client but it does not
carry much information about the machine and is hard to remember or
compare for humans. Nebraska supports an optional machinealias Omaha
field to display besides the machine ID.
Allow to set a MACHINE_ALIAS variable in the update.conf file to
specify the machine_alias field and fall back to the empty string
otherwise which will be ignored by Nebraska. As with the other
variables in /etc/flatcar/update.conf the MACHINE_ALIAS variable can be
changed while update-engine runs and it will pick up the change, which
is useful for scripts that set the machine alias like this:
  sudo sed -i "/MACHINE_ALIAS=.*/d" /etc/flatcar/update.conf
  echo "MACHINE_ALIAS=$(hostname)" | sudo tee -a /etc/flatcar/update.conf