src/update_engine_client: handle -update with idle
idle status was not correctly handled by the update_engine.

When a user wants to check for an update with `update_engine_client -update`
the command will fail with "Update failed" which can be misleading.
Actually, the update does not failed it is just cancelled because there
is no newer version.

Signed-off-by: Mathieu Tortuyaux <mathieu@kinvolk.io>