omaha_request_action_unittest: prevent init-order-fiasco with mock_system_state
Gmock registers actions in a global variable during mock object construction,
and mock_system_state is itself a global. This fails during program
initialization because initialization order of globals is undefined, and
results in a segfault. The mock_system_state is used when constructing
OmahaRequestParams, which are passed by value, so make the global function
static instead, and construct it on first use. This removes the segfault and
allows the tests to run.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>