workflow/c-cpp: use system gtest/gmock but custom bsdiff
Some of the PayloadProcessorTests depend on using a custom bspatch
implementation that allows specifying old/new extents. This version is
maintained in the chromiumos tree, so fetch from there and build. This
adds two new dependencies: libbrotli-dev libdivsufsort-dev

Building the tests also works against system gtest and gmock so we can
remove the step that builds those and install two packages.

The build also requires autoconf and libtool, and while the github
actions runners have those, a fresh ubuntu install doesn't. So add the
two deps for good measure.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>