Keep old nodes on cgroup v1 by default
In a coming release we will enable cgroupv2 by default in systemd. We are
trying to not break existing kubernetes deployments, which can't handle this
kind of change without intervention, so the postinstall hook will add the
'systemd.cgroup_unified_hierarchy' to the kernel command line to keep older
nodes on cgroupv1.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>