flatcar-postinst: Take the real path of the dynamic linker
When we find the dynamic linker in the new partition, then we derive
the library path by getting the basename of the dynamic linker. The
found linker and the library path are used to invoke binaries inside
the new partition (currently, cgpt).

On arm64, /lib and /lib64 are separate directories. Libc and the
dynamic linker are installed to /lib64, but there is also a symlink to
the dynamic linker in /lib. If the symlink in /lib gets found first,
then the library path is erroneously set to /lib. When calling the
found dynamic linker with the wrong library path, the linker will load
libraries from the old partition. This may work by accident, but
updating glibc to 2.34 broke it.