update_engine: Stop using deprecated curl options and info names
The CURLOPT_PROTOCOLS, CURLOPT_REDIR_PROTOCOLS options and the
CURLINFO_CONTENT_LENGTH_DOWNLOAD info name are deprecated, but
fortunately they have straightforward replacements.

Since we are running CI for update_engine on Ubuntu 22.04, which has
curl 7.81, we need to use some ifdeffery for CURLOPT_PROTOCOLS_STR and
CURLOPT_REDIR_PROTOCOLS_STR as these have been introduced in 7.85. On
Flatcar we have 7.87, so we will use the other branch of the ifdefs.

Such tricks were not needed for CURLINFO_CONTENT_LENGTH_DOWNLOAD as it
was deprecated since 7.55, so even the Ubuntu LTS has a newer version
of curl than that.