flatcar-postinst: Remove old VMware OEM workaround
In the past it was tried to support old installations by copying the
libs over that will be missing. We stopped doing that and instead
recommended reprovisioning or a helper update script. In the next weeks
we'll officially support an update path for OEM software.
Remove the old workaround that at the moment isn't used because it
wasn't updated to work with newer versions.