flatcar-postinst: In addition to SHA1, also check SHA256 hash for OEMs
The newer Omaha 3.1 hash_sha256 attribute is now supported by Nebraska
and should be used for OEM payloads. Up to now we only checked the
regular "hash" attribute for download integrity. It's not really
security critical because the payload has its own signature but it's
good to migrate all hashsum usage away from SHA1.
Find the hash and hash_sha256 attributes and require at least one to be
set for the OEM packages. Check the found hash attributes.