flatcar-postinst: Clean up incomplete move target
When the old OEM contents and a yet inactive systemd-sysext image are on
the OEM partition (or if no two sysexts fit on it), it can fail to write
the second systemd-sysext image on the OEM partition and the rootfs is
used instead. However, "mv" doesn't clean up the incomplete target file
which filled the disk, this results in the migration flag file creation
to fail, which fails the update action.
Clean up the incomplete move target that filled the disk.