flatcar-postinst: Ensure /etc/extensions is mergable
In Beta 3760.1.0 the /etc/extensions/ folder gets created by "mkdir -p"
because it does not exist in the lowerdir /usr/share/flatcar/etc/. This
causes the opaque marker to be set by overlayfs. The update to Alpha
thus does not merge the new /usr/share/flatcar/etc/extensions/ folder
with its docker/containerd sysext symlinks. We should have had
/etc/extensions/ in the lowerdir in Beta but didn't.

Ensure that the created folders are mergable by removing the overlayfs
marker. This is needed for existing installations and folders we expect
to exist in the lowerdir but were missing so far.