Only store Omaha XML response for update check
The full Omaha XML response is stored to a file for the postinst action,
but we only want the response for the initial update check passed this
way and not overwrite the file with responses for state reports.
Only store the Omaha XML response for the update check and not all of
the interaction with the Omaha server.