flatcar-postinst: Control target filename of extracted OEM payload
Nebraska servers with self-hosted synced packages use Omaha entries with
names like extrafile-amd64-3815.0.0-oem-ami.gz which can't be predicted.
The ue-rs matching option is not enough and we also need to control the
target filename because otherwise it would be
extrafile-amd64-3815.0.0-oem-ami.raw which we can't know either and
using shell globs is not the nicest solution.
Make use of two new ue-rs parameters to control the target filename and
make sure that we only use one match result.