Backport: Only store Omaha XML response for update check
This takes only the non-postinst action changes from
dfc4c523a320f45b8329a7ccc18123e063396f2e

The full Omaha XML response is stored to a file for the postinst action,
but we only want the response for the initial update check passed this
way and not overwrite the file with responses for state reports.
Only store the Omaha XML response for the update check and not all of
the interaction with the Omaha server.