flatcar-postinst: Support updating old airgapped clients
The XML passing and the passing of proxy env vars is required for
updating old airgapped clients because we need to support custom
non-public payload URLs and/or go through a proxy. However, these
changes are part of update-engine itself and thus are not present in old
clients. This blocks updating to the latest Stable.

Add two workarounds to read out proxy env vars from the service unit and
to read out the XML response from the journal logs.